Added Diesel Combustion Turbine to PEI & NL
Defined a new diesel combustion turbine technology,
"E_DIESEL-CT", in PEI and NL. This technology is
an affordable option for firm capacity. It is not
defined in NB and NS because these regions have
natural gas combustion turbines (NL and PEI do not
have NGCT due to an absence of a secure supply of NG).