Include RPS functionality and constraints
This commit makes the appropriate adjustments and additions
to the databases to permit the modelling of renewable portfolio
standards (RPS), clean electricity standards (CES), or similar
policy instruments.

Although it's possible to include an RPS in any of the model regions,
the policy is only included for NB and NS to reflect provincial
legislation.

Worth noting: Quebec imports to NB and NL imports to NS qualify for
the RPS. Imports from other jurisdictions do not.

To know exactly what qualifies/does  not qualify -- or to make changes
as you see fit -- consult the Efficiency table. Generators that produce
the ELCG-RPS commodity qualify whereas generators that produce the
ELCG commodity do not.