Change E_H2CC to E_H2NGCT
Prior to this commit, we defined a hydrogen combined cycle turbine (E_H2CC)
with the same specs as the natural gas combined cycle turbine, only it used
H2 as a fuel. This was based off a representation in the Open Energy Outlook.

We have updated that using data from the 2022 Nova Scotia Power Evergreen IRP.
This new technology, E_H2NGCT, is a combustion turbine that can use either
natural gas or H2.