Add emission intensity to imports.
A new emission commodity is added to the model to account
for emissions due to imported electricity. This emission
commodity, 'CO2e-Imports-Tax', is meant only to add a cost
premium to imports when there is a carbon tax.

At present, the actual emissions are not accounted for in
the importing region.